pyproject.toml
[project]
name = "component"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.18.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
